Move LayoutMain to layout folder
