+++ src/layout/LayoutMain/LayoutMain.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../Header/Header';
+import Header from '../../components/Header/Header';
 import styles from './LayoutMain.module.scss';
 
 const LayoutMain = ({ children }) => (
