+++ src/pages/about.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LayoutMain from '../components/LayoutMain/LayoutMain';
+import LayoutMain from '../layout/LayoutMain/LayoutMain';
 import About from '../components/About/About';
 import SEO from '../components/SEO/SEO';
